requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
flake8==7.1.1
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
python-dotenv==1.0.1
typing_extensions==4.12.2
waitress==3.0.0
Werkzeug==3.0.3
